Add automaton based implementation
I had a [solver implemented](http://gh.knutwalker.io/wordle-solvers) that is representing the known constraints
as an automaton and uses that one to search in an FST based on
a provided wordlist.

A comparison on my machine looks interesting

```
❯ hyperfine -w 1 -L implementation prune,cutoff,automaton -n '{implementation}' './target/release/roget --implementation {implementation}'
Benchmark 1: prune
  Time (mean ± σ):     17.646 s ±  0.055 s    [User: 17.525 s, System: 0.098 s]
  Range (min … max):   17.597 s … 17.785 s    10 runs

Benchmark 2: cutoff
  Time (mean ± σ):      8.708 s ±  0.005 s    [User: 8.653 s, System: 0.051 s]
  Range (min … max):    8.701 s …  8.718 s    10 runs

Benchmark 3: automaton
  Time (mean ± σ):     685.0 ms ±   1.1 ms    [User: 679.1 ms, System: 4.8 ms]
  Range (min … max):   683.3 ms … 686.7 ms    10 runs

Summary
  'automaton' ran
   12.71 ± 0.02 times faster than 'cutoff'
   25.76 ± 0.09 times faster than 'prune'
```

This implementation also uses a fixed first word, though that one can
also be calculated by using an empty automaton in the first iteration.