Parse dictionary at compile time
Interestingly enough, this brings a slight performance _decrease_ for
`cutoff`, `sigmoid`, and `escore` (~2-4%), but a performance _increase_
for `popular` (~15%). Even though strictly less work should have to
happen at runtime now.

This _does_ make it so that the dictionary is _always_ sorted by how
common a word is, but that shouldn't make a difference for most
implementations. Accuracy remains unchanged.