Merge all implementations
The equivalent configuration for `cache` here is about 2% slower than
the `cache` implementation when it was stand-alone.

The equivalent configuration for `popular` here is about 4% slower
than the `popular` implementation when it was stand-alone. _But_ it can
now be run with correctness caching enabled, which give it a 5.54x
speedup (!!).